Extract Quiz option button
